Add Blender and Rasterizer classes for rendering pipeline
This commit introduces two new classes, `Blender` and `Rasterizer`, to enhance the rendering pipeline in the Direct3D11 application. The `Blender` class manages blending states for transparency effects, while the `Rasterizer` class controls rasterization states, including backface culling options.

Shader files have been updated to support alpha masking and normal mapping, improving visual fidelity. Project files have also been modified to include the new source files and headers, ensuring proper compilation.

+++ Direct3D11Renderer/include/Core/Window.h
@@ -33,8 +33,8 @@ public:
 	Window& operator=(const Window&) = delete;
 	void SetTitle(const WCHAR* title);
 	static std::optional<int> ProcessMessages();
-	int GetWidth() const noexcept;
-	int GetHeight() const noexcept;
+	float GetWidth() const noexcept;
+	float GetHeight() const noexcept;
 
 	void CaptureMouse();
 	void ReleaseMouse();
@@ -50,8 +50,8 @@ private:
 	static LRESULT CALLBACK HandleMsgThunk(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam) noexcept;
 	LRESULT HandleMsg(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam) noexcept;
 
-	int width;
-	int height;
+	float width;
+	float height;
 	HWND hwnd;
 	bool mouseCaptured = false;
 	bool rawInputRegistered = false;
